+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add the `Cropper.create` static method for creating a new `Cropper` instance without the `new` operator.
+
 ## 1.6.3 (Aug 23, 2026)
 
 - Add `isFinite` check to prevent `NaN` in `getMaxZoomRatio` calculations.